Update README.md
Google doc was getting confusing, so I edited this README file for heading structure related to adding the mobile information. Also, there were edits in the Google doc for changing headings to title case and modifying terms like JavaScript or commonJS, so capitalization is correct. These changes are still indicated in the Google doc with yellow highlighted text and that Google doc is a but easier to read.